User auth

+++ flashcards/src/index.ts
@@ -83,6 +83,13 @@ resolver.define('updateFlashcard', async (req) => {
         };
     }
 
+    if (req.context.accountId && req.context.accountId != owner) {
+      return {
+        success: false,
+        error: "cannot edit someone else's flashcard"
+      }
+    }
+
     const updatedCard: Card = {
         ...existingCard,
         question_text: question_text || existingCard.question_text,
@@ -140,6 +147,13 @@ resolver.define('deleteFlashcard', async (req) => {
         success: false,
         error: `No card found with id: ${cardId}`,
       };
+    }
+
+    if (req.context.accountId && req.context.accountId != card.owner) {
+      return {
+        success: false,
+        error: "cannot delete someone else's flashcard"
+      }
     }
 
     await storage.delete(cardId);
@@ -265,6 +279,13 @@ resolver.define('updateDeck', async (req) => {
         };
     }
 
+    if (req.context.accountId && req.context.accountId != owner) {
+      return {
+        success: false,
+        error: "cannot edit someone else's deck"
+      }
+    }
+
     const updatedDeck: Deck = {
         ...existingDeck,
         title: title || existingDeck.title,
@@ -291,6 +312,13 @@ resolver.define('deleteDeck', async (req) => {
         success: false,
         error: `No deck found with id: ${deckId}`,
       };
+    }
+
+    if (req.context.accountId && req.context.accountId != deck.owner) {
+      return {
+        success: false,
+        error: "cannot delete someone else's deck"
+      }
     }
 
     await storage.delete(deckId);
@@ -350,6 +378,13 @@ resolver.define('addCardToDeck', async (req) => {
         };
     }
 
+    if (req.context.accountId && req.context.accountId != deck.owner) {
+      return {
+        success: false,
+        error: "cannot edit someone else's deck"
+      }
+    }
+
     deck.cards = [...(deck.cards || []), card];
 
     await storage.set(deckId, deck);
@@ -370,6 +405,13 @@ resolver.define('removeCardFromDeck', async (req) => {
             success: false,
             error: 'Item not found',
         };
+    }
+
+    if (req.context.accountId && req.context.accountId != deck.owner) {
+      return {
+        success: false,
+        error: "cannot edit someone else's deck"
+      }
     }
 
     deck.cards = deck.cards?.filter(c => c.id !== cardId) || [];
